Base/Build: Project Hellscream Core/Database 0.1.0a.

+++ src/server/game/Grids/Grid.h
@@ -1,20 +1,10 @@
-/*
- * Copyright (C) 2008-2016 TrinityCore <http://www.trinitycore.org/>
- * Copyright (C) 2005-2009 MaNGOS <http://getmangos.com/>
- *
- * This program is free software; you can redistribute it and/or modify it
- * under the terms of the GNU General Public License as published by the
- * Free Software Foundation; either version 2 of the License, or (at your
- * option) any later version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for
- * more details.
- *
- * You should have received a copy of the GNU General Public License along
- * with this program. If not, see <http://www.gnu.org/licenses/>.
- */
+////////////////////////////////////////////////////////////////////////////////
+//
+//  MILLENIUM-STUDIO
+//  Copyright 2016 Millenium-studio SARL
+//  All Rights Reserved.
+//
+////////////////////////////////////////////////////////////////////////////////
 
 #ifndef TRINITY_GRID_H
 #define TRINITY_GRID_H
@@ -33,6 +23,7 @@
 #include "Define.h"
 #include "TypeContainer.h"
 #include "TypeContainerVisitor.h"
+#include "Errors.h"
 
 // forward declaration
 template<class A, class T, class O> class GridLoader;
@@ -52,7 +43,7 @@ class Grid
         /** destructor to clean up its resources. This includes unloading the
         grid if it has not been unload.
         */
-        ~Grid() { }
+        ~Grid() {}
 
         /** an object of interested enters the grid
          */
@@ -104,7 +95,7 @@ class Grid
         template<class T>
         uint32 GetWorldObjectCountInGrid() const
         {
-            return uint32(i_objects.template Count<T>());
+            return i_objects.template Count<T>();
         }
 
         /** Inserts a container type object into the grid.
@@ -141,3 +132,4 @@ class Grid
         //ActiveGridObjects m_activeGridObjects;
 };
 #endif
+
